Fixed and rebuilt ScEd. Updated tex to handle greyscale PNGs.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@2610 3db68df2-c116-0410-a063-a993310a9797

+++ source/tools/sced/InfoBox.cpp
@@ -190,7 +190,7 @@ void CInfoBox::RenderInfo()
 		render(&fttext);
 		y-=dy;
 
-		u32 totalTris=m_LastStats.m_TerrainTris+m_LastStats.m_ModelTris;
+		u32 totalTris=(u32)(m_LastStats.m_TerrainTris+m_LastStats.m_ModelTris);
 		sprintf(buf,"TPF: %d",int(totalTris/m_LastStats.m_Counter));
 		COverlayText tpftext(5,y,0,DefaultFontName,buf,CColor(1,1,1,1));	
 		render(&tpftext);
